+++ srcs/parsing/parse_file.c
@@ -6,7 +6,7 @@
 /*   By: adbenoit <adbenoit@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/07 14:13:12 by adbenoit          #+#    #+#             */
-/*   Updated: 2020/12/20 23:52:13 by adbenoit         ###   ########.fr       */
+/*   Updated: 2020/12/21 04:10:02 by adbenoit         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 static int	parse_file(char **file, char *str, t_cmd **cmd, char **envp)
 {
-	size_t	i;
-	int		ret;
+	int	i;
+	int	ret;
 
 	i = 0;
 	if (*file)
